move simple files
